refactor: split up file so core can be used without napi bindings
